fix : Report 패키지 구조 변경  - admin, user로 소스코드 병합
